pass DataSource id to Feature
