+++ core/src/data/properties.cpp
@@ -1,15 +1,7 @@
-#include "tileData.h"
+#include "properties.h"
 #include <algorithm>
 
 namespace Tangram {
-
-int polygonSize(const Polygon& _polygon) {
-    int size = 0;
-    for (auto line : _polygon) {
-        size += line.size();
-    }
-    return size;
-}
 
 Properties::Properties(std::vector<Item>&& _items) {
     typedef std::vector<Item>::iterator iter_t;
@@ -51,21 +43,39 @@ const std::string& Properties::getString(const std::string& key) const {
     return EMPTY_STRING;
 }
 
-std::string Properties::getAsString(const std::string& key) const {
-
-    auto& it = get(key);
-    if (it.is<std::string>()) {
-        return it.get<std::string>();
-    } else if (it.is<int64_t>()) {
-        return std::to_string(it.get<int64_t>());
-    } else if (it.is<float>()) {
-        return std::to_string(it.get<float>());
+std::string Properties::asString(const Value& value) const {
+    if (value.is<std::string>()) {
+        return value.get<std::string>();
+    } else if (value.is<int64_t>()) {
+        return std::to_string(value.get<int64_t>());
+    } else if (value.is<float>()) {
+        return std::to_string(value.get<float>());
     }
     return "";
 }
 
+std::string Properties::getAsString(const std::string& key) const {
+
+    return asString(get(key));
+
+}
+
 void Properties::sort() {
     std::sort(props.begin(), props.end());
+}
+
+std::string Properties::toJson() const {
+
+    std::string json = "{ ";
+
+    for (const auto& item : props) {
+        bool last = (&item == &props.back());
+        json += "\"" + item.key + "\": \"" + asString(item.value) + (last ? "\"" : "\",");
+    }
+
+    json += " }";
+
+    return json;
 }
 
 }
