+++ core/src/util/geoJson.h
@@ -19,7 +19,7 @@ void extractPoly(const rapidjson::Value& _in, Polygon& _out, const Tile& _tile);
 
 void extractFeature(const rapidjson::Value& _in, Feature& _out, const Tile& _tile);
 
-void extractLayer(const rapidjson::Value& _in, Layer& _out, const Tile& _tile);
+void extractLayer(int32_t _sourceId, const rapidjson::Value& _in, Layer& _out, const Tile& _tile);
 
 }
